Reduction of network width by dividing the packets into smaller blocks using serializer and deserializer modules
Serializer and deserializer modules were created. This reduces the width of the network but increases its latency.

Either the number of blocks into which the packets are divided or the width of the smaller blocks is configurable.

A FIFO can optionally be added before each serializer module, to prevent messages being dropped because of the added latency.