Adding a compile time check for cache parameter conditions
As mentioned in the comments, cache parameters should satisfy the following conditions. 
The dcache supports multi-cycle fill/eviction with the following constraints:
- bank_width = block_width / assoc >= dword_width
- fill_width = N*bank_width <= block_width
But whenever an incorrect set of parameters are used, it will result in a compilation error(in verilator). 
To resolve this issue, one option is to edit the parameters such that it won't result in a compile error even when the conditions are not met. Other option is to introduce a compile time check of these conditions. 
This PR is for the second option, to communicate the cause of this error to the user, the above compile time check can be used. (Because generate blocks are resolved at compile time).